+++ utilities.h
@@ -397,7 +397,7 @@ gchar*			  getSequenceSegment(const char const *dnaSequence,
 
 const char*			   findFixedWidthFont(GtkWidget *widget);
 const char*			   findFixedWidthFontFamily(GtkWidget *widget, GList *pref_families);
-void                               getFontCharSize(GtkWidget *widget, PangoFontDescription *fontDesc, gint *width, gint *height);
+void                               getFontCharSize(GtkWidget *widget, PangoFontDescription *fontDesc, gdouble *width, gdouble *height);
 
 GtkWidget*                         createEmptyButtonBar(GtkToolbar **toolbar);
 void                               makeToolbarButton(GtkToolbar *toolbar, char *label, char *stockId, char *tooltip, GtkSignalFunc callback_func, gpointer data);
